requirements.txt
asgiref==3.2.10
bootstrap-admin==0.4.4
boto3==1.15.16
botocore==1.18.16
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
defusedxml==0.6.0
dj-database-url==0.5.0
dj-rest-auth==1.1.2
Django==3.1.2
django-allauth==0.43.0
django-cors-headers==3.5.0
django-environ==0.4.5
django-filter==2.4.0
django-heroku==0.3.1
-e git+https://github.com/jschneier/django-storages.git@738347ecfcf3e6354888f1100f7b336c729d5082#egg=django_storages
djangorestframework==3.12.1
gunicorn==20.0.4
idna==2.10
isodate==0.6.0
jmespath==0.10.0
msrest==0.6.19
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
stripe==2.55.1
urllib3==1.25.10
whitenoise==5.2.0
